refactor: 컴포넌트 경로 수정
